third_party/WebKit/LayoutTests/fullscreen/rendering/overflow-root.html

<!DOCTYPE html>
<title>overflow of root element when root element itself is fullscreen</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../trusted-event.js"></script>
<style>:root { overflow: scroll; }</style>
<div id="log"></div>
<script>
async_test(function()
{
    var root = document.documentElement;
    trusted_request(root, document.body);

    document.addEventListener("fullscreenchange", this.step_func_done(function()
    {
        // The scrollbar should remain.
        assert_equals(getComputedStyle(root).overflow, "scroll");
        assert_less_than(root.clientWidth, window.innerWidth);
    }));
});
</script>
